@bwplotka 2019-03-27
It looks quite nice but I need to spend more time on this & dive.. why :smile:  Thanks for this @brian-brazil !