@brian-brazil 2019-03-28
Did some further benchmarking and spotted a regression (not a bad one, but still), so made another tweak.

```
benchmark                                                            old ns/op      new ns/op      delta
BenchmarkHeadPostingForMatchers/n="1"-8                              3427           3415           -0.35%
BenchmarkHeadPostingForMatchers/n="1",j="foo"-8                      3842           4012           +4.42%
BenchmarkHeadPostingForMatchers/j="foo",n="1"-8                      2948           3139           +6.48%
BenchmarkHeadPostingForMatchers/n="1",j!="foo"-8                     5965           4560           -23.55%
BenchmarkHeadPostingForMatchers/i=~".*"-8                            4904256022     4903633195     -0.01%
BenchmarkHeadPostingForMatchers/i=~".+"-8                            5923029530     5763215376     -2.70%
BenchmarkHeadPostingForMatchers/i=~""-8                              1439518864     1441422425     +0.13%
BenchmarkHeadPostingForMatchers/i!=""-8                              6038586275     5837502229     -3.33%
BenchmarkHeadPostingForMatchers/n="1",i=~".*",j="foo"-8              71195312       64588907       -9.28%
BenchmarkHeadPostingForMatchers/n="1",i=~".*",i!="2",j="foo"-8       302829270      64726291       -78.63%
BenchmarkHeadPostingForMatchers/n="1",i!="",j="foo"-8                188138192      190979177      +1.51%
BenchmarkHeadPostingForMatchers/n="1",i=~".+",j="foo"-8              204630959      203860031      -0.38%
BenchmarkHeadPostingForMatchers/n="1",i=~"1.+",j="foo"-8             65171549       62710110       -3.78%
BenchmarkHeadPostingForMatchers/n="1",i=~".+",i!="2",j="foo"-8       420316223      203556233      -51.57%
BenchmarkHeadPostingForMatchers/n="1",i=~".+",i!~"2.*",j="foo"-8     354921692      344755168      -2.86%
```