@gouthamve 2019-04-08
Also, maximises the use of the shortcut for `labels.EqualMatcher` in `postingsForMatcher`.