@brian-brazil 2019-04-08
Yes, that's what it's doing.

On Mon 8 Apr 2019, 14:04 Goutham Veeramachaneni, <notifications@github.com>
wrote:

> Also, maximises the use of the shortcut for labels.EqualMatcher in
> postingsForMatcher.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/prometheus/tsdb/pull/572#issuecomment-480803689>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AGyTdsWmBtDDvOdqno1m9kuekkEVfg_rks5vezBngaJpZM4cOOIh>
> .
>