@brian-brazil 2019-04-08
I'm giving an example matcher, which matches empty. But also many non-empty
strings.

On Mon 8 Apr 2019, 16:01 Krasi Georgiev, <notifications@github.com> wrote:

> *@krasi-georgiev* commented on this pull request.
> ------------------------------
>
> In querier.go
> <https://github.com/prometheus/tsdb/pull/572#discussion_r273061708>:
>
> >
>  	for _, m := range ms {
> -		it, err := postingsForMatcher(ix, m)
> +		matchesEmpty := m.Matches("")
> +		if labelMustBeSet[m.Name()] || !matchesEmpty {
> +			// If this matcher must be non-empty, we can be smarter.
> +			nm, isNot := m.(*labels.NotMatcher)
> +			if isNot && matchesEmpty { // l!="foo"
>
> IIUC Brian is suggesting that this comment means that the condition is
> entered when the label value is empty which is not the case.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/prometheus/tsdb/pull/572#discussion_r273061708>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AGyTdig7u2nRgPSZNOl_udRblKUTxmL1ks5ve0uxgaJpZM4cOOIh>
> .
>